requirements.txt
gspread
google-auth
groq
python-dotenv
python-docx